+++ spec/openbox.spec.js
@@ -15,7 +15,7 @@ describe('The buying options section of the BBY API', function() {
         it('LIVE: Buying options search for all items', function(done) {
 
             // Product search for all items reviewed with exactly 4, show only name + sku
-            bby.buyingOptions(null, {
+            bby.openBox(null, {
                     show: 'name,sku'
                 })
                 .then(function(data) {
@@ -34,7 +34,7 @@ describe('The buying options section of the BBY API', function() {
         it('LIVE: Buying options search for one item', function(done) {
 
             // Product search for all items reviewed with exactly 4, show only name + sku
-            bby.buyingOptions(8610161)
+            bby.openBox(8610161)
                 .then(function(data) {
                     expect(data.results.length > 0).toBe(true);
                     expect(data.results[0].customerReviews.count).not.toBe(undefined);
@@ -52,7 +52,7 @@ describe('The buying options section of the BBY API', function() {
         it('LIVE: Buying options search for one item using callback', function(done) {
 
             // Product search for all items reviewed with exactly 4, show only name + sku
-            bby.buyingOptions(8610161, function(err, data) {
+            bby.openBox(8610161, function(err, data) {
                 expect(data.results.length > 0).toBe(true);
                 expect(data.results[0].customerReviews.count).not.toBe(undefined);
                 expect(data.results[0].customerReviews.averageScore).not.toBe(undefined);